requirements.txt
Flask==3.0.3
Werkzeug==3.0.1
pandas==2.2.2
openpyxl==3.1.2
jinja2==3.1.3
WeasyPrint==62.1
pdf2image==1.17.0
lxml==5.2.1
PyYAML==6.0.1

